refactor: split receipts core schema modules
